Added all of the remaining features and details
Since I've worked almost all day in this project, didnt commit as often as i should, kind of tired form all the work.

Added the searchBar, pokeModal and listItem components, modified the mainPage and the Home page to reflect the changes.

Added the image and the animations so it would look a little bit better!